Search/loader/requirements.txt
python-dotenv
azure-search-documents

requests